Conflicts:
	cocos/2d/CCActionGrid.h
	cocos/2d/CCActionGrid3D.h
	cocos/2d/CCActionInstant.h
	cocos/2d/CCActionInterval.h
	cocos/2d/CCActionPageTurn3D.h
	cocos/2d/CCActionProgressTimer.h
	cocos/2d/CCActionTiledGrid.h
	cocos/2d/CCActionTween.h
	cocos/2d/CCAnimation.h
	cocos/2d/CCAutoPolygon.h
	cocos/2d/CCCamera.h
	cocos/2d/CCClippingNode.h
	cocos/2d/CCClippingRectangleNode.h
	cocos/2d/CCDrawNode.h
	cocos/2d/CCFastTMXLayer.h
	cocos/2d/CCGrid.h
	cocos/2d/CCLabel.h
	cocos/2d/CCLayer.h
	cocos/2d/CCMenuItem.h
	cocos/2d/CCNode.h
	cocos/2d/CCParallaxNode.h
	cocos/2d/CCParticleSystem.h
	cocos/2d/CCParticleSystemQuad.h
	cocos/2d/CCProtectedNode.h
	cocos/2d/CCRenderTexture.h
	cocos/2d/CCSprite.h
	cocos/2d/CCSpriteFrameCache.h
	cocos/2d/CCTMXLayer.h
	cocos/2d/CCTMXTiledMap.h
	cocos/2d/CCTextFieldTTF.h
	cocos/3d/CCAABB.h
	cocos/3d/CCAnimate3D.h
	cocos/3d/CCAnimationCurve.h
	cocos/3d/CCBundleReader.h
	cocos/3d/CCMesh.h
	cocos/3d/CCOBB.h
	cocos/3d/CCSkeleton3D.h
	cocos/3d/CCSkybox.h
	cocos/3d/CCSprite3D.h
	cocos/3d/CCSprite3DMaterial.h
	cocos/3d/CCTerrain.h
	cocos/audio/include/AudioEngine.h
	cocos/base/CCAutoreleasePool.h
	cocos/base/CCData.h
	cocos/base/CCDirector.h
	cocos/base/CCEvent.h
	cocos/base/CCEventController.h
	cocos/base/CCEventDispatcher.h
	cocos/base/CCEventListener.h
	cocos/base/CCIMEDispatcher.h
	cocos/base/CCProtocols.h
	cocos/base/CCRef.h
	cocos/base/CCScriptSupport.h
	cocos/base/CCTouch.h
	cocos/base/CCVector.h
	cocos/base/ccConfig.h
	cocos/editor-support/cocostudio/ActionTimeline/CCActionTimeline.h
	cocos/editor-support/cocostudio/CCActionFrame.h
	cocos/editor-support/cocostudio/CCActionManagerEx.h
	cocos/editor-support/cocostudio/CCActionNode.h
	cocos/editor-support/cocostudio/CCArmatureAnimation.h
	cocos/editor-support/cocostudio/CCArmatureDataManager.h
	cocos/editor-support/cocostudio/CCComAttribute.h
	cocos/editor-support/cocostudio/CCComAudio.h
	cocos/editor-support/cocostudio/CCComController.h
	cocos/editor-support/cocostudio/CCComRender.h
	cocos/math/CCAffineTransform.h
	cocos/math/Mat4.h
	cocos/network/CCDownloader.h
	cocos/network/HttpClient.h
	cocos/network/HttpRequest.h
	cocos/network/HttpResponse.h
	cocos/network/SocketIO.h
	cocos/network/WebSocket.h
	cocos/physics/CCPhysicsBody.h
	cocos/physics/CCPhysicsContact.h
	cocos/physics/CCPhysicsHelper.h
	cocos/physics/CCPhysicsJoint.h
	cocos/physics/CCPhysicsShape.h
	cocos/physics/CCPhysicsWorld.h
	cocos/physics3d/CCPhysics3DConstraint.h
	cocos/physics3d/CCPhysics3DObject.h
	cocos/physics3d/CCPhysics3DShape.h
	cocos/platform/CCFileUtils.h
	cocos/platform/CCImage.h
	cocos/renderer/CCGLProgramState.h
	cocos/renderer/CCGroupCommand.h
	cocos/renderer/CCMaterial.h
	cocos/renderer/CCPrimitive.h
	cocos/renderer/CCPrimitiveCommand.h
	cocos/renderer/CCRenderCommand.h
	cocos/renderer/CCRenderState.h
	cocos/renderer/CCRenderer.h
	cocos/renderer/CCTexture2D.h
	cocos/renderer/CCTextureCache.h
	cocos/renderer/CCTrianglesCommand.h
	cocos/renderer/CCVertexIndexBuffer.h
	cocos/renderer/CCVertexIndexData.h
	cocos/scripting/js-bindings/manual/ScriptingCore.h
	cocos/scripting/lua-bindings/manual/CCLuaBridge.h
	cocos/scripting/lua-bindings/manual/CCLuaEngine.h
	cocos/scripting/lua-bindings/manual/CCLuaStack.h
	cocos/scripting/lua-bindings/manual/CCLuaValue.h
	cocos/scripting/lua-bindings/manual/LuaBasicConversions.h
	cocos/scripting/lua-bindings/manual/cocos2d/LuaOpengl.h
	cocos/scripting/lua-bindings/manual/cocos2d/LuaScriptHandlerMgr.h
	cocos/scripting/lua-bindings/manual/cocosbuilder/CCBProxy.h
	cocos/scripting/lua-bindings/manual/platform/android/CCLuaJavaBridge.h
	cocos/scripting/lua-bindings/manual/tolua_fix.h
	cocos/ui/UIEditBox/UIEditBox.h
	cocos/ui/UIHelper.h
	cocos/ui/UILayout.h
	cocos/ui/UILayoutComponent.h
	cocos/ui/UIListView.h
	cocos/ui/UIPageView.h
	cocos/ui/UIScale9Sprite.h
	cocos/ui/UIScrollView.h
	cocos/ui/UIText.h
	cocos/ui/UITextField.h
	cocos/ui/UIVideoPlayer.h
	cocos/ui/UIWebView.h
	cocos/ui/UIWidget.h
	extensions/assets-manager/AssetsManager.h
	extensions/assets-manager/AssetsManagerEx.h
	extensions/assets-manager/Manifest.h
	plugin
	tests/cpp-tests/Resources/ccs-res
	tools/bindings-generator
	tools/cocos2d-console
	web

+++ cocos/2d/CCFontFreeType.h
@@ -53,7 +53,7 @@ class CC_DLL FontFreeType : public Font
 public:
     static const int DistanceMapSpread;
 
-    static FontFreeType* create(const std::string &fontName, int fontSize, GlyphCollection glyphs, 
+    static FontFreeType* create(const std::string &fontName, float fontSize, GlyphCollection glyphs,
         const char *customGlyphs,bool distanceFieldEnabled = false,int outline = 0);
 
     static void shutdownFreeType();
@@ -83,7 +83,7 @@ private:
     FontFreeType(bool distanceFieldEnabled = false, int outline = 0);
     virtual ~FontFreeType();
 
-    bool createFontObject(const std::string &fontName, int fontSize);
+    bool createFontObject(const std::string &fontName, float fontSize);
 
     bool initFreeType();
     FT_Library getFTLibrary();
